tst/requirements.txt
ecs_logging
lxml
peewee
pyXMLSecurity
python-keycloak
requests
